Refactor monolithic messageCreateTranslate event
