+++ events/translation/limits.ts
@@ -1,0 +1,2 @@
+export const DISCORD_MESSAGE_CHARS_LIMIT = 2000;
+export const DISCORD_ATTACHMENT_SIZE_LIMIT = 25 * 1024 * 1024; // 25 MB
